забыл чет

+++ src/main/java/find/itTeam/dto/CreateTag.java
@@ -3,5 +3,5 @@ import lombok.Data;
 @Data
 public class CreateTag {
     private String tag;
-    private int developer_id;
+    private Long developer_id;
 }
